+++ src/Recyclarr.Cli/Console/Helpers/CliLogLevel.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel;
+
+namespace Recyclarr.Cli.Console.Helpers;
+
+[TypeConverter(typeof(KebabCaseEnumTypeConverter<CliLogLevel>))]
+internal enum CliLogLevel
+{
+    Debug,
+    Info,
+    Warn,
+}