+++ src/Recyclarr.Cli/Console/Helpers/KebabCaseEnumTypeConverter.cs
@@ -11,6 +11,11 @@ internal sealed class KebabCaseEnumTypeConverter<TEnum> : TypeConverter
         return sourceType == typeof(string) || base.CanConvertFrom(context, sourceType);
     }
 
+    public override bool CanConvertTo(ITypeDescriptorContext? context, Type? destinationType)
+    {
+        return destinationType == typeof(string) || base.CanConvertTo(context, destinationType);
+    }
+
     public override object? ConvertFrom(
         ITypeDescriptorContext? context,
         CultureInfo? culture,